backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
minio==7.2.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.2.0
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2 